Transfer create... to TableBuilder and del Creator
