+++ test/test_tablebuilder.py
@@ -1,7 +1,7 @@
 import pytest
 
 from PdfVisualisation.TableStyle import TableStyle
-from model.tables.Creator import Creator
+from model.tables.TableBuilder import TableBuilder
 
 
 @pytest.fixture()
@@ -10,6 +10,5 @@ def table_data():
     return ["f", "s", "t"], [1, 2, 3], table_style.normal
 
 
-def test_creator_call(table_data):
-    creator = Creator()
-    creator.create_fixedwidth_table(table_data[0], table_data[1], table_data[2])
+def test_tablebuilder_fixedwidth_call(table_data):
+    TableBuilder.create_fixedwidth_table(table_data[0], table_data[1], table_data[2])
